Unredactor/requirements.txt
Flask==3.1.2
PyMuPDF==1.26.7
pandas==2.3.3
python-docx==1.2.0